Treat Benchmark::ArgName like Benchmark::Arg
Bug: #1257
Test: TODO

Benchmark::Arg creates an instance of the Benchmark which accepts a
single argument.

Similarly, Benchmark::Args creates an instance of the Benchmark which
accepts multiple arguments.

Benchmark::ArgNames works like ::Args - it must match the number of
arguments passed to Args, and it matches each Name to an Arg.

Previously, within a family of Benchmarks there was no way to name the
Arg(s) of the second (and so on) instances differently from the first
instance. Add this feature by treating ArgName like Arg - convert
Benchmark::arg_names_ to a vector of vectors (like args_). Each inner
vector corresponds to an instance.

Possible regressions:
1) If a client was depending on ArgNames to name the arguments for every
instance in a family, it will now only name the first instance. This
will still be supported by calling ArgNames multiple times with the same
values.
    - Alternatively, the library could support this use by using the
    last set of ArgNames for any further instances.
2) Previously, a client could call ArgNames or ArgName multiple times on
the same Benchmark, and only the last one would be respected. I don't
expect this to happen in practice.

Change-Id: Ide3f8cb433e219042a3506c192e8c4287ae55780