Fix build error on Clang 13 on Windows
By default, MSVC's `stddef.h` header defines an `offsetof` macro containing a `reinterpret_cast`. This breaks building `benchmark.cc` on Clang 13 on Windows, because a `reinterpret_cast` is not allowed within a `static_assert`. Defining the `_CRT_USE_BUILTIN_OFFSETOF` uses a builtin for `offsetof` instead of the macro, which allows the file to build again.