Fix failing perf gtests
This is more of a *tentative* pull request to understand better the
architecture, as I think the changes in this commit are more
profound than they should be for just a "test fix".

I cannot understand why the PerfCounters counters_ object inside
PerfCounterMeasurement was made static. There is really no reason
for that from a perf_event point of view. I am here thinking that
the reason for the static is some higher level process/thread
issue that I do not familiar with. If so, kindly disregard this
pull request.

Once you make that object non-static, the entire architecture
simplifies a lot. I created a shared pointer on the off-chance that
the static had to do with copy/move semantics and the way the top
level class was created.

I could not create that shared pointer with std::make_shared
because the relevant PerfCounters constructor is private and
std::forward is obviously not a friend class. So we got an extra
int malloc there. But I suspect this will have no impact since
these objects are so infrequently created.

I did not want to replace the shared_ptr with a unique_ptr since
the gain in performance would be really nil and the danger of
introducing some move semantics glitch, much higher.

The nice thing is that this shared pointer also acts as a
std::optional - in the sense you can query it through its bool
operator.

All tests were passing then - apart from the multithreaded in
release mode. I had to make sure it stayed there computing for
an arbitrary 0.25 seconds until the tests passed. Apparently it
was going too fast - or even perhaps the compiler was
optimizing away the entire squared sum. It is passing
consistently now. I changed a bit the name of the variables
there so I could understand better. Hope you do not mind.

If the basics of this pull request are sound, there is more to
clean up like do away with the NoCounters() method which only
returns an empty shared_ptr at the moment.

Please let me know. Thank you.