Propose PMCs shared from the top
The proposed change makes the creation and sharing of performance
counters local at RunBenchmarks and then shared through constructor
with all the benchmarks.

The performance counters are now created as an object and shared even
if empty. Because of this change, the method IsValid() now really does not
have any reason to be and perhaps should be cleaned up. The current test
as it is, only checks if the size of the names vector is the same as the size
of the counter ids. This will be unnecessary with the "filter" patch that
 was just approved yesterday. The tests were changed to reflect this
 new meaning.

I feel the approach proposed is not pointless refactoring. It will make
counters more visible  than sharing from the bottom and will lead to
easier maintenance and collaboration going forward.

All tests are passing, with and without libpfm. Didnt' test on Windows
but I guess we're about to know.