Simplify assembly tests
I would like to propose the following changes in the assembly tests with the goal to simplify their maintenance.

The changes are local and should not affect any other tests or code.

The first change under this proposal is moving the underlying tooling from COMPILER+PYTHON to OBJDUMP+SED. The advantages are:
First, objdump's generated assembly is standard no matter what compiler is being used - because it is coming back from object file. This will simplify support to any future compiler we decide to adopt.
Second, the "cleaning" step is not necessary anymore since we do not have to deal with each compiler's idiosyncrasies. From objdump's disassembly it is pretty much a couple of simple regex rules to clean addresses and offsets.
Third, the way we were generating assembly code by injecting an "-S" flag to the end of the compiler's command line is supported on clang and gcc only. Granted, these are the only supported compilers but in the current proposal we do not need to touch on the compiler's command line at all. By using objdump we will be future-proofing the process both in terms of new versions and new compilers.
Third, by using SED we can aggregate all the regex rules under a very neat SED script under `/tools/strip_asm.sed`, which is very easy to understand and maintain.

The second change is to normalize all instruction and register names. Instead of matching `%eax` we will be matching REG. Instead of matching -0xc(%rip) we will be matching OFFSET(RIP). This normalization comes in handy as we unlock another level of complexity. For example, the GCC compiler generates `inc %eax` while clang generates `add 1,%eax` and nvidia hpc writes `add 1,0xc(%rdi)`. We can match all these three versions with a single rule `CHECK: {{inc |add 1,}}{{REG|OFFSET(REG}}}` instead of creating three separate sets of rules for each compiler. This is seen in `test/donotoptimize_assembly_test.cc:179`.

The third change is to add levels of detail to the check rules. We can now check for all the compilers with `CHECK:`, check for a specific compiler with `CHECK-CLANG:` or `CHECK-GNU:`, check for a major version with `CHECK-CLANG-12:` and even check for major and minor releases as `CHECK-CLANG-5_0`. This allows flexibility in creating exceptions to the rule.

As this is a proof of concept at this point, I have modified only the donotoptimize_assembly_test.cc test, there is still clobber_memory_assembly_test.cc and state_assembly_test.cc to rewrite the rules.

Thank you!