Allow for collection of more PMUs than are physically available.
Normally we can only collect as many PMUs as we have physical PMU
counters because they are managed as a single pinned group and extra
requested PMUs will not be collected. With this patch we instead collect
each PMU as a separate group once this limit is hit using PMU
multiplexing.